get_active_partition move to HAL
