+++ spm/spm_hal_sw.c
@@ -15,6 +15,7 @@
 #include "cmsis_os2.h"
 #include "psa_defs.h"
 #include "spm_internal.h"
+#include "spm_panic.h"
 
 #if SPM_LEVEL == 0
 
@@ -47,3 +48,23 @@ void memory_protection_init(const mem_region_t *regions, uint32_t region_count)
 }
 
 #endif // SPM_LEVEL == 0
+
+#if (SPM_LEVEL == 0) || (SPM_LEVEL == 1)
+
+extern spm_db_t g_spm;
+
+spm_partition_t *get_active_partition(void)
+{
+    osThreadId_t active_thread_id = osThreadGetId();
+    SPM_ASSERT(NULL != active_thread_id);
+
+    for (uint32_t i = 0; i < g_spm.partition_count; ++i) {
+        if (g_spm.partitions[i].thread_id == active_thread_id) {
+            return &(g_spm.partitions[i]);
+        }
+    }
+
+    return NULL; // Valid in case of NSPE
+}
+
+#endif // (SPM_LEVEL == 0) || (SPM_LEVEL == 1)
